Expose more test scaffolding for use in implementing libraries
